Build: make `kotlinx.interval` a module part of a multi-project setup
