+++ kotlinx.interval/build.gradle.kts
@@ -1,0 +1,12 @@
+plugins {
+    id( "interval.library-conventions" )
+}
+
+publishing {
+    publications.filterIsInstance<MavenPublication>().forEach {
+        it.pom {
+            name.set("kotlinx.interval")
+            description.set("Kotlin multiplatform bounded open/closed generic intervals.")
+        }
+    }
+}
