+++ kotlinx.interval/src/commonMain/kotlin/BasicTypeIntervals.kt
