+++ kotlinx.interval/src/commonMain/kotlin/BasicTypeOperations.kt
