+++ kotlinx.interval/src/commonMain/kotlin/Interval.kt
