+++ kotlinx.interval/src/commonMain/kotlin/IntervalTypeOperations.kt
