+++ kotlinx.interval/src/commonMain/kotlin/TypeOperations.kt
