+++ kotlinx.interval/src/commonTest/kotlin/BasicTypeIntervalsTest.kt
