+++ kotlinx.interval/src/commonTest/kotlin/BasicTypeOperationsTest.kt
