+++ kotlinx.interval/src/commonTest/kotlin/IntervalTest.kt
