+++ kotlinx.interval/src/commonTest/kotlin/TypeOperationsTest.kt
