+++ settings.gradle.kts
@@ -1,1 +1,3 @@
 rootProject.name = "kotlinx-interval"
+
+include( "kotlinx.interval" )
